+++ src/functions/external.ts
@@ -7,7 +7,7 @@ const plugin = (external: Set<string>): Plugin => ({
   name: "external",
   setup({ onResolve }) {
     onResolve({ filter: packageNameRegex, namespace: "file" }, args => {
-      external.add(args.path);
+      external.add(packageNameRegex.exec(args.path)![0]);
       return { path: args.path, external: true };
     });
   },
@@ -15,20 +15,15 @@ const plugin = (external: Set<string>): Plugin => ({
 
 /**
  * Bundle a file to find out which libraries it is REALLY depending on.
- * (unused imports are dropped by esbuild.)
+ * The file's author should guarantee that the file can be bundled.
  * @example
- * external("a.ts")
+ * external("a.ts", { loader: { ".css": "text" } })
  * // a.ts: import "esbuild"
  * //=> ["esbuild"]
+ * NOTE: it only outputs DIRECT dependencies, not indirect ones.
  */
 export async function external(entry: string, options?: BuildOptions) {
-  options = {
-    entryPoints: [entry],
-    bundle: true,
-    format: "esm",
-    write: false,
-    ...options,
-  };
+  options = { entryPoints: [entry], bundle: true, format: "esm", write: false, ...options };
   const libraries = new Set<string>();
   (options.plugins ??= []).push(plugin(libraries));
   await esbuild.build(options);