+++ test/index.ts
@@ -1,5 +1,5 @@
 export const a = 42;
-import { external } from "../src";
+import { external, platform } from "../src";
 export default 1;
 
 console.log("import.meta =", import.meta);
@@ -8,5 +8,10 @@ if (require.main) {
   console.log("__filename =", __filename);
 } else {
   console.log("require =", require);
-  external("./test/external.ts").then(console.log);
+  external("./test/external.ts").then(e => {
+    console.log("test external():", e);
+  });
+  platform("./test/external.ts").then(e => {
+    console.log("test platform():", e);
+  });
 }